Use double precision in JitterDemo

+++ src/JitterDemo/Demos/Demo20.cs
@@ -32,8 +32,8 @@ public struct CollisionTriangle : ISupportMappable
 
     public void SupportMap(in JVector direction, out JVector result)
     {
-        float min = JVector.Dot(A, direction);
-        float dot = JVector.Dot(B, direction);
+        double min = JVector.Dot(A, direction);
+        double dot = JVector.Dot(B, direction);
 
         result = A;
         if (dot > min)
@@ -94,7 +94,7 @@ public class CustomCollisionDetection : IBroadPhaseFilter
             ts.C = octree.Vertices[octree.Indices[index].IndexC];
 
             bool hit = NarrowPhase.MPREPA(ts, rbs, rbs.RigidBody!.Orientation, rbs.RigidBody!.Position,
-                out JVector pointA, out JVector pointB, out _, out float penetration);
+                out JVector pointA, out JVector pointB, out _, out double penetration);
 
             if (hit)
             {
@@ -129,9 +129,9 @@ public class Tester(Octree tree) : IDynamicTreeProxy, IRayCastable
     public JVector Velocity => JVector.Zero;
     public JBBox WorldBoundingBox { get; } = tree.Dimensions;
 
-    public bool RayCast(in JVector origin, in JVector direction, out JVector normal, out float lambda)
+    public bool RayCast(in JVector origin, in JVector direction, out JVector normal, out double lambda)
     {
-        return tree.Raycast(origin,direction, out normal, out lambda);
+        return tree.Raycast(origin, direction, out normal, out lambda);
     }
 }
